fixed latex.
needed to rename artwork.png figure from duplicate nat to artwork

needed to move lfig in artwork figure after caption. see stefaanlippens.net/latexcaptionlabel/

The order of \caption and \label declarations matter in LaTeX floats (tables and figures). It is important to get references to the float right. The \label should come after the \caption or even inside the \caption environment. If you put the \label before the \caption you will get a reference to the (sub)section where float is declared instead of a reference to the float itself.