used more beautiful code for the same bug
The bug was that the server didn't receive a register message, though
the connection was established to the host client (after informing the
host client that the server was up, the host client spinned in a new
thread until the connection was established and then send the register
message. this was apparently too fast because ESCH was not yet ready
serverside)
Solution was first to sleep for 10ms between that.
Fix is now that the first Serverside ClientHandler (ESCH) informs the
ServerActivityCallbackInterface, in our case the LobbyList activity,
when it has acquired the readlock for the socket and is ready to
communicate.

The LobbyListActivity then tells the client to send a register message.

Not sure why this is neccessary.