connecting now also works on the hardware devices
the problem was that the server was not yet ready when the client sent the register message.
easy fix: wait 1s before sending that (after calling connect())

reason why this worked on emulators is probably that the client is faster relative to the network when running on real devices... but why would the network be slow on the emulators?