moved some settings to Debug.java and Globals.java
I'm trying to fix the bug of not connecting if the register message is
sent too soon on real devices in wifi. According to my logic, making the
server take longer to initialize after the connection is established
should increase the chance that the register message gets lost. However,
even setting the serverside sleep to 30 seconds and the clientside sleep
completely off, it still works on the emulators...

This must either mean that the emulators buffer the network traffic for
me, or that my reasoning is broken. I need to try this on real
devices...

NOTE: currently, Debug.CLIENT_SLEEP_AFTER_CONNECTION_ESTABLISHED is set
to false and Debug.SERVER_SLEEP_AFTER_CONNECTION_ESTABLISHED  to 30'000,
so I'd expect this commit to fail to register on the real devices