still not understanding how a cast from long to float could cause this though
I ran two samples now, the difference because of the cast can be in a
magnitude of 16000 ms. So a huge difference was still rounded to 0.0 as
float. but why did it then always stop with time passed being 131072 ms?
that's 2^17. is that maybe the next larger float possible for numbers of
this size?