fixed bug that made it impossible to remove something from player's hand.
Reason was: we were using player.hand.remove(card), but it would be possibly causing inconsistencies if we were to override Card.equals() and Card.hashCode(). The current solution is a new method in Player which only compares the cards by ID, not by any other attributes. So a card with different origin and parameters would still be removed if it happens to be the first hit.